+++ CHANGELOG.md
@@ -6,9 +6,16 @@
 
 ### Minor Changes
 
-- Added endpoints for `isAirdropNft()` and `reportSpam()` on the nft namespace.
+## 3.1.1
+
+### Major Changes
+
+### Minor Changes
+
+- Added endpoints for `isAirdropNft()` and `reportSpam()` in `NftNamespace`.
 - Added a missing `marketplaceAddress` field to the `NftSale` in `NftNamespace.getNftSales()`.
 - Fixed an incorrect field name for `bannerImageUrl` in `OpenSeaCollectionMetadata`. Deprecated the old field name `imageBannerUrl`.
+- Added support for Base networks in `NotifyNamespace`.
 
 ## 3.1.0